core/merge: Mark documents for deletion
  Until now, when we detected the deletion of a document either on the
  local file system or the remote Cozy, we would merge it by deleting
  it from the PouchDB database.
  Then, The Sync module would detect that change via the PouchDB changes
  feed and propagate it.

  However, deleted documents in PouchDB cannot be queried by `get()`.
  Also, when multiple modifications are made to the same document (i.e.
  saved with the same `_id` attribute), the changes feed will return
  only the latest version of the document.

  Those 2 "restrictions" make merging some changes harder. For example:
  - the remote addition of files or folders into a locally deleted
    folder (i.e. the parent folder can't be queried so it's "missing")
  - the deletion and creation of a file at the same path/_id (i.e. we
    lose the _rev of the deleted document and can't propagate it as an
    update)

  The solution we found to this problem is to mark documents for
  deletion during the Merge step so they can still be queried until the
  deletion has been propagated. This way we can decide to cancel the
  deletion or take any other appropriate action to merge other changes
  that could not be merged (or merged in a way that would result in an
  incorrect or impossible propagation).

  This commit introduces the `deleted` marker and does not make use of
  it to solve our problems yet. Its goal is to achieve feature parity.

+++ core/merge.js
@@ -103,7 +103,7 @@ class Merge {
 
     try {
       const folder = await this.pouch.db.get(parentId)
-      if (folder) {
+      if (folder && !folder.deleted) {
         return
       }
     } catch (err) {
@@ -163,6 +163,10 @@ class Merge {
     metadata.markSide(side, doc, file)
     metadata.assignMaxDate(doc, file)
     if (file) {
+      if (file.deleted) {
+        return this.updateFileAsync(side, doc)
+      }
+
       const idConflict /*: ?IdConflictInfo */ = IdConflict.detect(
         { side, doc },
         file
@@ -261,7 +265,7 @@ class Merge {
         if (doc.mime == null) {
           doc.mime = file.mime
         }
-      } else if (!metadata.isAtLeastUpToDate(side, file)) {
+      } else if (!file.deleted && !metadata.isAtLeastUpToDate(side, file)) {
         if (side === 'local') {
           // We have a merged but unsynced remote update and we can't create a
           // conflict because the local rename will trigger an overwrite of the
@@ -388,6 +392,10 @@ class Merge {
 
       const file /*: ?Metadata */ = await this.pouch.byIdMaybeAsync(doc._id)
       if (file) {
+        if (file.deleted) {
+          doc.overwrite = file
+        }
+
         const idConflict /*: ?IdConflictInfo */ = IdConflict.detect(
           { side, doc, was },
           file
@@ -462,6 +470,10 @@ class Merge {
 
     const folder /*: ?Metadata */ = await this.pouch.byIdMaybeAsync(doc._id)
     if (folder) {
+      if (folder.deleted) {
+        doc.overwrite = folder
+      }
+
       const idConflict /*: ?IdConflictInfo */ = IdConflict.detect(
         { side, doc, was },
         folder
@@ -547,8 +559,18 @@ class Merge {
       bulk.push(src)
 
       const existingDstRev = existingDstRevs[dst._id]
+      // Filtering out deleted destination docs would mean failing to save the new version.
+      // However, replacing the deleted docs will mean failing to propagate the change.
       if (existingDstRev && folder.overwrite) {
-        dst._rev = existingDstRev
+        if (src.deleted) {
+          // Don't create destination that would overwrite existing file
+          continue
+          // TODO: document with test deleting child file then move its parent
+          // folder, overwriting another folder with child file of same name.
+          // The file in the destination folder should not be deleted.
+        } else {
+          dst._rev = existingDstRev
+        }
       }
 
       // FIXME: Find a cleaner way to pass the syncPath to the Merge
@@ -631,7 +653,7 @@ class Merge {
     const oldMetadata /*: ?Metadata */ = await this.pouch.byIdMaybeAsync(
       was._id
     )
-    if (!oldMetadata) {
+    if (!oldMetadata || oldMetadata.deleted) {
       log.debug({ path }, 'Nothing to trash')
       return
     }
@@ -658,7 +680,7 @@ class Merge {
     newMetadata.trashed = true
     if (oldMetadata.sides && oldMetadata.sides[side]) {
       metadata.markSide(side, oldMetadata, oldMetadata)
-      oldMetadata._deleted = true
+      oldMetadata.deleted = true
       try {
         await this.pouch.put(oldMetadata)
         return
@@ -689,7 +711,11 @@ class Merge {
     let children = await this.pouch.byRecursivePathAsync(was._id)
     children = children.reverse()
     for (let child of Array.from(children)) {
-      if (child.docType === 'file' && !metadata.isUpToDate(side, child)) {
+      if (
+        child.docType === 'file' &&
+        !child.deleted &&
+        !metadata.isUpToDate(side, child)
+      ) {
         delete was.trashed
         delete was.errors
         if (was.sides) {
@@ -699,14 +725,18 @@ class Merge {
           // information.
           was.sides = { target: 1, [otherSide(side)]: 1 }
         }
+        // TODO: why prevent removing all files that were up-to-date?
         return this.putFolderAsync(otherSide(side), was)
       }
     }
     // Remove in pouchdb the sub-folders
+    //
+    // TODO: We could only one loop if the update of one child would not prevent
+    // the trashing of the other children.
     for (let child of Array.from(children)) {
       if (child.docType === 'folder') {
         try {
-          child._deleted = true
+          child.deleted = true
           await this.pouch.put(child)
         } catch (err) {
           log.warn({ path, err })
@@ -724,10 +754,10 @@ class Merge {
   async deleteFileAsync(side /*: SideName */, doc /*: Metadata */) {
     log.debug({ path: doc.path }, 'deleteFileAsync')
     const file /*: ?Metadata */ = await this.pouch.byIdMaybeAsync(doc._id)
-    if (!file) return null
+    if (!file || file.deleted) return null
     if (file.sides && file.sides[side]) {
       metadata.markSide(side, file, file)
-      file._deleted = true
+      file.deleted = true
       delete file.errors
       return this.pouch.put(file)
     } else {
@@ -746,7 +776,7 @@ class Merge {
   async deleteFolderAsync(side /*: SideName */, doc /*: Metadata */) {
     log.debug({ path: doc.path }, 'deleteFolderAsync')
     const folder /*: ?Metadata */ = await this.pouch.byIdMaybeAsync(doc._id)
-    if (!folder) return null
+    if (!folder || folder.deleted) return null
     if (folder.sides && folder.sides[side]) {
       return this.deleteFolderRecursivelyAsync(side, folder)
     } else {
@@ -766,7 +796,9 @@ class Merge {
     docs = docs.reverse()
     docs.push(folder)
     const toPreserve = new Set()
-    for (let doc of Array.from(docs)) {
+    for (let doc of docs) {
+      if (doc.deleted) continue
+
       if (
         toPreserve.has(doc.path) ||
         (doc.sides && !metadata.isUpToDate(side, doc))
@@ -785,7 +817,7 @@ class Merge {
         toPreserve.add(path.dirname(doc.path))
       } else {
         metadata.markSide(side, doc, doc)
-        doc._deleted = true
+        doc.deleted = true
         delete doc.errors
       }
     }
